GVT-3009: Separate SearchDropdown into its own component and allow narrowing the search by asset type
