+++ ui/src/track-layout/track-layout-search-api.ts
@@ -8,6 +8,7 @@ import {
 import { getNonNull, queryParams } from 'api/api-fetch';
 import { TRACK_LAYOUT_URI, contextInUri } from 'track-layout/track-layout-api';
 import { LayoutContext } from 'common/common-model';
+import { SearchType } from 'tool-bar/search-dropdown';
 
 export interface LayoutSearchResult {
     switches: LayoutSwitch[];
@@ -19,6 +20,7 @@ export interface LayoutSearchResult {
 export async function getBySearchTerm(
     searchTerm: string,
     layoutContext: LayoutContext,
+    types: SearchType[],
     locationTrackSearchScope?: LocationTrackId,
     limitPerResultType: number = 10,
 ): Promise<LayoutSearchResult> {
@@ -28,6 +30,7 @@ export async function getBySearchTerm(
         searchTerm: searchTerm,
         locationTrackSearchScope: locationTrackSearchScope,
         limitPerResultType: limitPerResultType,
+        types: types,
     });
 
     return await getNonNull<LayoutSearchResult>(`${uri}${params}`);
